pyproject.toml
[tool.poetry]
name = "ai_safety_leaderboard"
version = "0.1.0"
description = "A short description of the project"
authors = ["Kirill Duvakin <duvakin.k@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.9"

[tool.poetry.group.frontend.dependencies]
gradio = "^4.42.0"
requests = "^2.32.3"
gradio-leaderboard = "^0.0.11"
pandas = "^2.2.2"

[tool.poetry.group.backend.dependencies]
fastapi = "^0.112.3"
uvicorn = "^0.30.6"
garak = "^0.9.0.15"
psycopg = "^3.2.1"
pydantic-settings = "^2.4.0"
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pre-commit = "*"

[tool.poetry.scripts]
run_service = "src:main"

[tool.nbqa.mutate]
black = 1
autoflake = 1
isort = 1

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"